feat(agent): Added Commandcode. Loc error.
